examples/otel/requirements.txt
datalayer-core
fastapi
uvicorn
pydantic
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
